refactor(problem): create problem class and make SiteProblem inherit
To support work to extend out to the RoutingProblem class, shared attributes and methods have been moved into a new _Problem private class, with SiteProblem also being tidied up to move attribute initialisation methods and related display methods into a mixin
